% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment2somites.R
\name{segment2somites}
\alias{segment2somites}
\title{Infer individual somite period and length}
\usage{
segment2somites(df, delta_threshold = 50, x = "time", y = "length")
}
\arguments{
\item{df}{A data-frame with length for whole segmented region per embryo.}

\item{delta_threshold}{A numeric vector, with the "jump" in length when a new somite is formed.}

\item{x}{A numeric vector, indicating the time.}

\item{y}{A numeric vector, indicating the length.}
}
\value{
A data-frame with period and length for individual somites.
}
\description{
This function takes as input consecutive measurements of the whole
segmented region from a vertebrate embryo and it finds the individual
somites.
Using an heuristic based on consecutive differences between lengths,
a length difference between two consecutive points greater than the
delta_threshold (default is 50.0 micrometers) is used to delimit a somite.
}
