% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somite_periods_doc.R
\docType{data}
\name{somite_periods}
\alias{somite_periods}
\title{Embryo Somite Periods data}
\format{
A \code{data-frame} with 9 variables:
\describe{
\item{\code{embryo_id}}{Categorical variable identifying the embryo. Thirteen embryos were measured (EmbryoA-EmbryoM). Note: EmbryoJ and EmbryoK were separately measured for the early and late somites, labeled EmbryoJ1, EmbryoJ2, EmbryoK1, and EmbryoK2.}
\item{\code{somite_id}}{Categorial variable indicating the somite number.}
\item{\code{DevelTime}}{Categorical variable, indicating Early or Late development. Early: somites 01-09. Late: somites 14-20.}
\item{\code{start_time}}{Time of the first frame where the new somite appears in the global development (in minutes).}
\item{\code{somite_position_mean}}{Mean of the whole segmented region length (in micrometers) for all frames where each somite is visible (from the first frame where somite_n appears, to the last frame before somite_n+1 appears.}
\item{\code{somite_position_sd}}{Standard deviation of somite_position_mean.}
\item{\code{somite_period}}{Individual somite period, in minutes, inferred from the raw segmented length data.}
\item{\code{somite_length}}{Individual somite length, in micrometers (individual somite size).}
\item{\code{somite_length_sd}}{Standard deviation for individual somite length.}
}

For further details contact Isabel Duarte (\email{iduarte.scientist@gmail.com}).
}
\usage{
somite_periods
}
\description{
\describe{
\item{\bold{DATA DESCRIPTION:}}{
Somite periods inferred from the segmented region length.
Visit the help page for the embryo_lengths dataset for further details about
the experimental setting.
This dataset contains the results from the automated inference of individual
somite length and period using only the raw segmented region length data
collected by the researchers.}
}
}
\keyword{datasets}
